+++ Backend/Accounts/Accounts.Core/Entities/UserRole.cs
@@ -6,7 +6,8 @@ namespace Accounts.Core
     {
         public string ProjectId { get; set; }
         public string RoleId { get; set; }
-        public long UserId { get; set; }
+        public int UserId { get; set; }
+
         public string CreatedBy { get; set; }
         public DateTime CreatedAt { get; set; }
