+++ Backend/Accounts/Accounts.Core/Repositories/IUserRepository.cs
@@ -2,7 +2,7 @@
 
 namespace Accounts.Core
 {
-    public interface IUserRepository<TUser>  : IRepository<TUser, long> where TUser : User
+    public interface IUserRepository<TUser>  : IRepository<TUser, int> where TUser : User
     {
     }
 }
