+++ Backend/Accounts/Accounts.Data/Repositories/UserRepository.cs
@@ -4,7 +4,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Accounts.Data
 {
-    public class UserRepository<TUser>: Repository<TUser, long>, IUserRepository<TUser> where TUser : User
+    public class UserRepository<TUser>: Repository<TUser, int>, IUserRepository<TUser> where TUser : User
     {
         public UserRepository(AccountDbContext context) : base(context)
         {
