+++ Backend/Accounts/Accounts.Data/Repositories/UserRoleRepository.cs
@@ -6,18 +6,18 @@ using System.Linq;
 
 namespace Accounts.Data
 {
-    public class UserRoleRepository : Repository<UserRole, long, string, string>, IUserRoleRepository
+    public class UserRoleRepository : Repository<UserRole, int, string, string>, IUserRoleRepository
     {
         public UserRoleRepository(AccountDbContext context) : base(context)
         {
         }
 
-        public IList<UserRole> GetAllUserRole(string projectId, long userId)
+        public IList<UserRole> GetAllUserRole(string projectId, int userId)
         {
             return Entities.Include(x => x.User).Include(z => z.Role).Where(x => x.ProjectId == projectId && x.UserId == userId).ToList();
         }
 
-        public UserRole GetUserRole(string projectId, string roleId, long userId)
+        public UserRole GetUserRole(string projectId, string roleId, int userId)
         {
             return Entities.Include(x => x.User).Include(z => z.Role).Where(x => x.ProjectId == projectId && x.RoleId == roleId && x.UserId == userId).FirstOrDefault();
         }
